pyproject.toml
[tool.poetry]
name = "NINA"
version = "0.1.0b"
description = "This is a project that provides both a place for people to seek information about their favorite dolls. It just happens to also have a Battle Royale Simulation Utility."
license = "EPL-2.0"
authors = ["Tech. TTGames <50541739+Tech-TTGames@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/Tech-TTGames/NINA"
packages = [
    {include = "NINA"}
]

[tool.poetry.scripts]
start = "NINA.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"
colorama = "^0.4.6"
discord-py = "^2.6.3"
aiohttp = {extras = ["speedups"], version = "^3.12.14"}
pillow = "^11.3.0"
text-to-owo = "^2.0.1"
tomli-w = { version = "^1.2.0", optional = true }
beautifulsoup4 = { version = "^4.14.2", optional = true }
selenium = {version = "^4.35.0", optional = true}
lxml = {version = "^6.0.2", optional = true}
mkdocs-material = "^9.6.21"
mkdocs-git-revision-date-localized-plugin = "^1.3.0"
mkdocs-git-committers-plugin-2 = "^2.5.0"
mkdocs-awesome-pages-plugin = "^2.10.1"
cairosvg = "^2.8.2"
markdown-grid-tables = "^0.6.0"
mkdocs-glightbox = "^0.5.1"
mkdocs-ezlinked-plugin = "^0.3.3"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
yapf = "^0.43.0"
pylint = "^3.3.8"

[tool.poetry.extras]
brentutils = ["tomli-w", "beautifulsoup4", "selenium", "lxml"]

[tool.yapf]
based_on_style = "google"
column_limit = 120

[tool.isort]
profile = "google"
lines_between_types = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
